requirement-mac.txt
gym==0.26.2
imageio==2.36.0
matplotlib==3.9.2
numpy==2.1.3
pygame==2.6.1
python-dotenv==1.0.1
Requests==2.32.3

--pre torch torchvision --index-url https://download.pytorch.org/whl/nightly/cpu